+++ electron/wine-manager/types.ts
