+++ electron/wine-manager/utils.ts
